put the applicationContextIdFilter in correct context
