+++ spring-boot-project/spring-boot-actuator-autoconfigure/src/main/java/org/springframework/boot/actuate/autoconfigure/web/servlet/ServletManagementContextAutoConfiguration.java
@@ -21,11 +21,8 @@ import javax.servlet.Servlet;
 import org.springframework.boot.actuate.autoconfigure.endpoint.web.WebEndpointProperties;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication.Type;
-import org.springframework.boot.web.servlet.filter.ApplicationContextHeaderFilter;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -45,23 +42,10 @@ public class ServletManagementContextAutoConfiguration {
 	public ServletManagementContextFactory servletWebChildContextFactory() {
 		return new ServletManagementContextFactory();
 	}
-
 	@Bean
 	public ManagementServletContext managementServletContext(WebEndpointProperties properties) {
 		return properties::getBasePath;
 	}
-
-	// Put Servlets and Filters in their own nested class so they don't force early
-	// instantiation of ManagementServerProperties.
-	@Configuration(proxyBeanMethods = false)
-	@ConditionalOnProperty(prefix = "management.server", name = "add-application-context-header", havingValue = "true")
-	protected static class ApplicationContextFilterConfiguration {
-
-		@Bean
-		public ApplicationContextHeaderFilter applicationContextIdFilter(ApplicationContext context) {
-			return new ApplicationContextHeaderFilter(context);
-		}
-
-	}
+	//       git remote set-url [--push] <name> <newurl> [<oldurl>]
 
 }
